fix: CR fixes
