+++ src/components/Avatar/Avatar.jsx
@@ -1,10 +1,33 @@
-import { useRef } from 'react'
+import { useRef, useState } from 'react'
+import axios from 'axios'
+import { toast } from 'react-toastify'
 import UserImage from '/src/components/UserImage'
 import * as Styled from './Avatar.styled'
 
-const Avatar = ({ user, onUpdateAvatar, imageKey }) => {
+const Avatar = ({ user }) => {
 
   const fileInput = useRef(null)
+  const [imageKey, setImageKey] = useState(null)
+
+  const onUpdateAvatar = async (file) => {
+    try {
+      const user_name = user.name
+      const imageKey = `?${Math.random().toString(36).substring(2, 15)}-${Date.now()}`
+      const opts = {
+        headers: {
+          'Content-Type': file.type,
+        },
+      }
+      await axios.put(`/api/users/${user_name}/avatar`, file, opts)
+      toast.success('Profile updated')
+      setImageKey(imageKey)
+    } catch (error) {
+      console.log(error)
+      toast.error('Unable to update avatar')
+      toast.error(error.details)
+    }
+  }
+
   const handleInputChange = (e) => {
     e.preventDefault()
     if (!e.target.files || !e.target.files[0]) return
@@ -18,9 +41,6 @@ const Avatar = ({ user, onUpdateAvatar, imageKey }) => {
           <Styled.AvatarIcon onClick={() => fileInput.current.click()} icon='edit' />
           <UserImage size={100} name={user.name} imageKey={imageKey} />
         </Styled.ImageIcon>
-        <Styled.Username>
-          {user?.attrib?.fullName}
-        </Styled.Username>
     </Styled.Avatar>
   )
 }
