+++ src/services/user/getUsers.js
@@ -206,6 +206,5 @@ export const {
   useLazyGetUserQuery,
   useGetUsersAssigneeQuery,
   useGetMeQuery,
-  useGetUserAvatarQuery,
   useGetUserSessionsQuery,
 } = getUsers
